requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
ecdsa==0.19.2
et_xmlfile==2.0.0
fastapi==0.135.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mlxtend==0.24.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.42.0
Werkzeug==3.1.6
